Convert feet to inches for length

+++ priv/repo/migrations/20180119040116_create_surfboards.exs
@@ -4,7 +4,7 @@ defmodule Partywave.Repo.Migrations.CreateSurfboards do
   def change do
     create table(:surfboards) do
       add :model, :string
-      add :length, :float
+      add :length, :integer
       add :width, :float
       add :thickness, :float
       add :volume, :float
